skills/handsomestwei/patent-disclosure-skill/requirements.txt
# 使用 tools/md_to_docx.py、docx_to_md.py、pptx_to_md.py 时安装
python-docx>=1.1.0
# 定稿图示（mermaid → PNG）须 Node：在 tools/ 执行 npm install 或 npx，见 tools/README.md
mammoth>=1.6.0
python-pptx>=0.6.21
